Move blog list to own component
